move kse quantizer to separate dir
