+++ src/sdk/pynni/nni/compressors/torch_compressor/__init__.py
@@ -1,3 +1,4 @@
 from .pruner import LevelPruner, AGPruner, SensitivityPruner
-from .quantizer import NaiveQuantizer, DoReFaQuantizer, QATquantizer, KSE
+from .quantizer import NaiveQuantizer, DoReFaQuantizer, QATquantizer
+from .kse_quantizer.kse_quantizer import KSEQuantizer
 from ._nnimc_torch import TorchPruner, TorchQuantizer
