+++ src/sdk/pynni/nni/compressors/torch_compressor/quantizer.py
@@ -2,7 +2,6 @@ import torch
 import torch.nn as nn
 from ._nnimc_torch import TorchQuantizer
 from ._nnimc_torch import _torch_default_get_configure, _torch_default_load_configure_file, _torch_detect_module
-from .Conv2d_KSE import Conv2d_KSE
 import logging
 logger = logging.getLogger('torch quantizer')
 
@@ -110,66 +109,3 @@ class QATquantizer(TorchQuantizer):
         out = out*scale + zero_point
         orig_type = weight.dtype
         return out.type(orig_type)
-
-
-class KSE(TorchQuantizer):
-    """
-    Use algorithm from "ExploitingKernelSparsityandEntropyforInterpretableCNNCompression" 
-    https://arxiv.org/abs/1812.04368
-    """
-    def __init__(self, configure_list):
-        """
-            configure Args:
-                G: 
-                T: 
-        """
-        super().__init__()
-        self.configure_list = []
-        if isinstance(configure_list, list):
-            for configure in configure_list:
-                self.configure_list.append(configure)
-        else:
-            raise ValueError('please init with configure list')
-
-    def get_GT(self, configure):
-        if not isinstance(configure, dict):
-            logger.warning('WARNING: you should input a dict to get_GT, set DEFAULT { }')
-            configure = {}
-        G = configure.get('G', 4)
-        T = configure.get('T', 0)
-        return G, T
-    
-    def compress(self, model):
-        super().compress(model)
-        for layer_info in _torch_detect_module(model, nn.Conv2d):
-            G, T = self.get_GT(_torch_default_get_configure(self.configure_list, layer_info))
-
-            # replace origin conv2d with conv2d_kse
-            conv2d_kse = Conv2d_KSE(
-                input_channels = layer_info.layer.in_channels, 
-                output_channels = layer_info.layer.out_channels, 
-                kernel_size = layer_info.layer.kernel_size, 
-                stride = layer_info.layer.stride, 
-                padding = layer_info.layer.padding, 
-                bias=False, 
-                G=G, 
-                T=T)
-            conv2d_kse.weight = layer_info.layer.weight
-            conv2d_kse.bias = layer_info.layer.bias
-
-            # calculate clusters and index
-            conv2d_kse.KSE(G, T)
-            conv2d_kse.forward_init()
-            layer_info.layer = conv2d_kse
-            setattr(model, layer_info.name, layer_info.layer)
-            self._instrument_layer(layer_info)
-    
-    def _instrument_layer(self, layer_info):
-        assert layer_info._forward is None
-        layer_info._forward = layer_info.layer.forward
-
-        def new_forward(*input):
-            # layer_info.layer.weight.data = self.quantize_weight(layer_info, layer_info.layer.weight.data)
-            return layer_info._forward(*input)
-
-        layer_info.layer.forward = new_forward
