+++ src/sdk/pynni/tests/test_compressor.py
@@ -113,7 +113,7 @@ class CompressorTestCase(TestCase):
     def test_torch_KSE_quantizer(self):
         model = TorchMnist()
         configure_list = [{'G':4, 'T':0, 'support_type':'default'}]
-        torch_compressor.KSE(configure_list).compress(model)
+        torch_compressor.KSEQuantizer(configure_list).compress(model)
 
 if __name__ == '__main__':
     main()
